requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
aiofiles==23.2.1
jinja2==3.1.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pillow==10.1.0
ffmpeg-python==0.2.0
boto3==1.34.0
minio==7.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
email-validator==2.1.0.post1